feat(cli): extract cli defenitions into seperate modules
