+++ src/cli/commands/mod.rs
@@ -1,0 +1,4 @@
+pub mod create;
+pub mod delete;
+pub mod rename;
+pub mod list;
